:card_file_box: Truncate table in tests

+++ modules/drift-canary/src/test/scala/attribution/health/VerificationResult.scala
@@ -11,7 +11,7 @@ import java.time.LocalDate
 final case class VerificationResult(
     date: LocalDate,
     expectedCount: Int,
-    actualCount: Double,
+    actualCount: Int,
     minimumSuccessThreshold: Double,
     successRate: Double,
     status: VerificationResult.Status,
